+++ src/felupe/region/_region.py
@@ -310,6 +310,12 @@ class Region:
         region = self
 
         if mesh is not None:
+
+            if "container" in type(mesh).__name__.lower():
+                raise TypeError(
+                    "A mesh container is not supported by a region, use a mesh instead."
+                )
+
             region.mesh = mesh
 
         if element is not None: